+++ app/Http/Controllers/Admin/RoleController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Http\Requests\App\StoreRoleRequest;
 use App\Http\Requests\App\UpdateRoleRequest;
 use App\Models\Role;
